+++ CHANGELOG.md
@@ -37,5 +37,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   bit length is not a multiple of 8.
 - Fixed `esp:deep_sleep/1` that did not accept values above 31 minutes.
 - Fixed a bug that could cause processes to hang indefinitely when calling ports that have terminated.
+- Fixed potential VM crash when parsing external terms.
 
 ## [0.5.0] - 2022-03-22